backend/requirements.txt
pandas>=2.2.0
numpy>=2.0
tableauhyperapi==0.0.24457
openpyxl==3.1.5
lxml==4.9.3
requests==2.31.0
python-dateutil==2.8.2
sqlalchemy 
python-dotenv
fastapi
uvicorn
requests
openai
Pillow
numpy
anthropic
python-multipart
jspdf